Made console output easier to quickly scan
For me it was very hard to scroll through the console output with a large amount of tests and find what I was looking for. These changes make several things easier to scan: when detailed reporting is on, you can very quickly find the start of any given set of tests, and you can very quickly find the summary of any set of tests, and you can quickly scan through the tests themselves to zero in on which ones have failed.  Works well for me.